+++ DatadogInternal/Sources/NetworkInstrumentation/B3/B3HTTPHeadersWriter.swift
@@ -59,6 +59,9 @@ public class B3HTTPHeadersWriter: TracePropagationHeadersWriter {
 
     private let samplingStrategy: TraceSamplingStrategy
 
+    /// Defines whether the trace context should be injected into all requests or only sampled ones.
+    private let traceContextInjection: TraceContextInjection
+
     /// The telemetry header encoding used by the writer.
     private let injectEncoding: InjectEncoding
 
@@ -85,7 +88,8 @@ public class B3HTTPHeadersWriter: TracePropagationHeadersWriter {
     ) {
         self.init(
             samplingStrategy: .custom(sampleRate: sampleRate),
-            injectEncoding: injectEncoding
+            injectEncoding: injectEncoding,
+            traceContextInjection: .all
         )
     }
 
@@ -93,12 +97,15 @@ public class B3HTTPHeadersWriter: TracePropagationHeadersWriter {
     ///
     /// - Parameter samplingStrategy: The strategy for sampling trace propagation headers.
     /// - Parameter injectEncoding: The B3 header encoding type, with `.single` as the default.
+    /// - Parameter traceContextInjection: The trace context injection strategy, with `.all` as the default.
     public init(
         samplingStrategy: TraceSamplingStrategy,
-        injectEncoding: InjectEncoding = .single
+        injectEncoding: InjectEncoding = .single,
+        traceContextInjection: TraceContextInjection = .all
     ) {
         self.samplingStrategy = samplingStrategy
         self.injectEncoding = injectEncoding
+        self.traceContextInjection = traceContextInjection
     }
 
     /// Writes the trace ID, span ID, and optional parent span ID into the trace propagation headers.
@@ -112,30 +119,35 @@ public class B3HTTPHeadersWriter: TracePropagationHeadersWriter {
 
         typealias Constants = B3HTTPHeaders.Constants
 
-        switch injectEncoding {
-        case .multiple:
-            traceHeaderFields = [
-                B3HTTPHeaders.Multiple.sampledField: sampled ? Constants.sampledValue : Constants.unsampledValue
-            ]
-
-            if sampled {
-                traceHeaderFields[B3HTTPHeaders.Multiple.traceIDField] = String(traceContext.traceID, representation: .hexadecimal32Chars)
-                traceHeaderFields[B3HTTPHeaders.Multiple.spanIDField] = String(traceContext.spanID, representation: .hexadecimal16Chars)
-                traceHeaderFields[B3HTTPHeaders.Multiple.parentSpanIDField] = traceContext.parentSpanID.map { String($0, representation: .hexadecimal16Chars) }
-            }
-        case .single:
-            if sampled {
-                traceHeaderFields[B3HTTPHeaders.Single.b3Field] = [
-                    String(traceContext.traceID, representation: .hexadecimal32Chars),
-                    String(traceContext.spanID, representation: .hexadecimal16Chars),
-                    sampled ? Constants.sampledValue : Constants.unsampledValue,
-                    traceContext.parentSpanID.map { String($0, representation: .hexadecimal16Chars) }
+        switch (traceContextInjection, sampled) {
+        case (.all, _), (.sampled, true):
+            switch injectEncoding {
+            case .multiple:
+                traceHeaderFields = [
+                    B3HTTPHeaders.Multiple.sampledField: sampled ? Constants.sampledValue : Constants.unsampledValue
                 ]
-                .compactMap { $0 }
-                .joined(separator: Constants.b3Separator)
-            } else {
-                traceHeaderFields[B3HTTPHeaders.Single.b3Field] = Constants.unsampledValue
+
+                if sampled {
+                    traceHeaderFields[B3HTTPHeaders.Multiple.traceIDField] = String(traceContext.traceID, representation: .hexadecimal32Chars)
+                    traceHeaderFields[B3HTTPHeaders.Multiple.spanIDField] = String(traceContext.spanID, representation: .hexadecimal16Chars)
+                    traceHeaderFields[B3HTTPHeaders.Multiple.parentSpanIDField] = traceContext.parentSpanID.map { String($0, representation: .hexadecimal16Chars) }
+                }
+            case .single:
+                if sampled {
+                    traceHeaderFields[B3HTTPHeaders.Single.b3Field] = [
+                        String(traceContext.traceID, representation: .hexadecimal32Chars),
+                        String(traceContext.spanID, representation: .hexadecimal16Chars),
+                        sampled ? Constants.sampledValue : Constants.unsampledValue,
+                        traceContext.parentSpanID.map { String($0, representation: .hexadecimal16Chars) }
+                    ]
+                    .compactMap { $0 }
+                    .joined(separator: Constants.b3Separator)
+                } else {
+                    traceHeaderFields[B3HTTPHeaders.Single.b3Field] = Constants.unsampledValue
+                }
             }
+        case (.sampled, false):
+            break
         }
     }
 }