+++ DatadogInternal/Tests/NetworkInstrumentation/B3HTTPHeadersReaderTests.swift
@@ -81,7 +81,7 @@ class B3HTTPHeadersReaderTests: XCTestCase {
 
     func testReadingSampledTraceContext() {
         let encoding: B3HTTPHeadersWriter.InjectEncoding = [.multiple, .single].randomElement()!
-        let writer = B3HTTPHeadersWriter(samplingStrategy: .custom(sampleRate: 100), injectEncoding: encoding)
+        let writer = B3HTTPHeadersWriter(samplingStrategy: .custom(sampleRate: 100), injectEncoding: encoding, traceContextInjection: .all)
         writer.write(traceContext: .mockRandom())
 
         let reader = B3HTTPHeadersReader(httpHeaderFields: writer.traceHeaderFields)
@@ -89,13 +89,23 @@ class B3HTTPHeadersReaderTests: XCTestCase {
         XCTAssertEqual(reader.sampled, true)
     }
 
-    func testReadingNotSampledTraceContext() {
+    func testReadingNotSampledTraceContext_givenTraceContextInjectionIsAll() {
         let encoding: B3HTTPHeadersWriter.InjectEncoding = [.multiple, .single].randomElement()!
-        let writer = B3HTTPHeadersWriter(samplingStrategy: .custom(sampleRate: 0), injectEncoding: encoding)
+        let writer = B3HTTPHeadersWriter(samplingStrategy: .custom(sampleRate: 0), injectEncoding: encoding, traceContextInjection: .all)
         writer.write(traceContext: .mockRandom())
 
         let reader = B3HTTPHeadersReader(httpHeaderFields: writer.traceHeaderFields)
         XCTAssertNil(reader.read(), "When not sampled, it should return no trace context")
         XCTAssertEqual(reader.sampled, false)
+    }
+
+    func testReadingNotSampledTraceContext_givenTraceContextInjectionIsSampled() {
+        let encoding: B3HTTPHeadersWriter.InjectEncoding = [.multiple, .single].randomElement()!
+        let writer = B3HTTPHeadersWriter(samplingStrategy: .custom(sampleRate: 0), injectEncoding: encoding, traceContextInjection: .sampled)
+        writer.write(traceContext: .mockRandom())
+
+        let reader = B3HTTPHeadersReader(httpHeaderFields: writer.traceHeaderFields)
+        XCTAssertNil(reader.read(), "When not sampled, it should return no trace context")
+        XCTAssertNil(reader.sampled)
     }
 }